fix review suggestions
