+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/dao/constants/SQLConstants.java
@@ -3019,43 +3019,67 @@ public class SQLConstants {
                     "AM_GW_API_ARTIFACTS.REVISION_ID=AM_GW_API_DEPLOYMENTS.REVISION_ID";
 
     public static final String RETRIEVE_ALL_ARTIFACTS_BY_DATA_PLANE_IDS =
-            "SELECT AM_GW_API_DEPLOYMENTS.API_ID AS API_ID,AM_GW_API_DEPLOYMENTS.REVISION_ID AS REVISION_ID," +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_PROVIDER AS API_PROVIDER," +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_NAME AS API_NAME,AM_GW_PUBLISHED_API_DETAILS.API_VERSION AS API_VERSION," +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_TYPE AS API_TYPE,AM_GW_API_ARTIFACTS.ARTIFACT AS ARTIFACT," +
-                    "AM_GW_API_DEPLOYMENTS.LABEL AS LABEL,AM_GW_API_DEPLOYMENTS.VHOST AS VHOST, " +
-                    "CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID AS DATA_PLANE_ID, " +
-                    "AM_API.CONTEXT AS CONTEXT FROM " +
-                    "AM_GW_PUBLISHED_API_DETAILS,AM_GW_API_ARTIFACTS,AM_GW_API_DEPLOYMENTS,AM_API," +
-                    "CHOREO_GW_ENV_DATA_PLANE_MAPPING,AM_GW_VHOST,AM_GATEWAY_ENVIRONMENT WHERE " +
-                    "CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID IN (_DATA_PLANE_IDS_) AND " +
-                    "AM_GW_VHOST.GATEWAY_ENV_ID=AM_GATEWAY_ENVIRONMENT.ID AND " +
-                    "AM_GATEWAY_ENVIRONMENT.UUID=CHOREO_GW_ENV_DATA_PLANE_MAPPING.GATEWAY_ENV_UUID AND " +
-                    "AM_GW_VHOST.HOST=AM_GW_API_DEPLOYMENTS.VHOST AND " +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND " +
-                    "AM_GW_API_ARTIFACTS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND " +
-                    "AM_API.API_UUID=AM_GW_API_DEPLOYMENTS.API_ID AND " +
-                    "AM_GW_API_ARTIFACTS.REVISION_ID=AM_GW_API_DEPLOYMENTS.REVISION_ID";
+            "SELECT " +
+                    "   AM_GW_API_DEPLOYMENTS.API_ID AS API_ID," +
+                    "   AM_GW_API_DEPLOYMENTS.REVISION_ID AS REVISION_ID," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_PROVIDER AS API_PROVIDER," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_NAME AS API_NAME," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_VERSION AS API_VERSION," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_TYPE AS API_TYPE," +
+                    "   AM_GW_API_ARTIFACTS.ARTIFACT AS ARTIFACT," +
+                    "   AM_GW_API_DEPLOYMENTS.LABEL AS LABEL," +
+                    "   AM_GW_API_DEPLOYMENTS.VHOST AS VHOST," +
+                    "   CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID AS DATA_PLANE_ID," +
+                    "   AM_API.CONTEXT AS CONTEXT " +
+                    "FROM" +
+                    "   AM_GW_PUBLISHED_API_DETAILS," +
+                    "   AM_GW_API_ARTIFACTS," +
+                    "   AM_GW_API_DEPLOYMENTS," +
+                    "   AM_API," +
+                    "   CHOREO_GW_ENV_DATA_PLANE_MAPPING," +
+                    "   AM_GW_VHOST," +
+                    "   AM_GATEWAY_ENVIRONMENT " +
+                    "WHERE" +
+                    "   CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID IN (_DATA_PLANE_IDS_) AND" +
+                    "   AM_GW_VHOST.GATEWAY_ENV_ID=AM_GATEWAY_ENVIRONMENT.ID AND" +
+                    "   AM_GATEWAY_ENVIRONMENT.UUID=CHOREO_GW_ENV_DATA_PLANE_MAPPING.GATEWAY_ENV_UUID AND" +
+                    "   AM_GW_VHOST.HOST=AM_GW_API_DEPLOYMENTS.VHOST AND" +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND" +
+                    "   AM_GW_API_ARTIFACTS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND" +
+                    "   AM_API.API_UUID=AM_GW_API_DEPLOYMENTS.API_ID AND" +
+                    "   AM_GW_API_ARTIFACTS.REVISION_ID=AM_GW_API_DEPLOYMENTS.REVISION_ID";
 
     public static final String RETRIEVE_ALL_ARTIFACTS_BY_ORGANIZATION_AND_DATA_PLANE_IDS =
-            "SELECT AM_GW_API_DEPLOYMENTS.API_ID AS API_ID,AM_GW_API_DEPLOYMENTS.REVISION_ID AS REVISION_ID," +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_PROVIDER AS API_PROVIDER," +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_NAME AS API_NAME,AM_GW_PUBLISHED_API_DETAILS.API_VERSION AS API_VERSION," +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_TYPE AS API_TYPE,AM_GW_API_ARTIFACTS.ARTIFACT AS ARTIFACT," +
-                    "AM_GW_API_DEPLOYMENTS.LABEL AS LABEL,AM_GW_API_DEPLOYMENTS.VHOST AS VHOST, " +
-                    "CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID AS DATA_PLANE_ID, " +
-                    "AM_API.CONTEXT AS CONTEXT FROM " +
-                    "AM_GW_PUBLISHED_API_DETAILS,AM_GW_API_ARTIFACTS,AM_GW_API_DEPLOYMENTS,AM_API," +
-                    "CHOREO_GW_ENV_DATA_PLANE_MAPPING,AM_GW_VHOST,AM_GATEWAY_ENVIRONMENT WHERE " +
-                    "CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID IN (_DATA_PLANE_IDS_) AND " +
-                    "AM_GW_PUBLISHED_API_DETAILS.TENANT_DOMAIN = ? AND " +
-                    "AM_GW_VHOST.GATEWAY_ENV_ID=AM_GATEWAY_ENVIRONMENT.ID AND " +
-                    "AM_GATEWAY_ENVIRONMENT.UUID=CHOREO_GW_ENV_DATA_PLANE_MAPPING.GATEWAY_ENV_UUID AND " +
-                    "AM_GW_VHOST.HOST=AM_GW_API_DEPLOYMENTS.VHOST AND " +
-                    "AM_GW_PUBLISHED_API_DETAILS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND " +
-                    "AM_GW_API_ARTIFACTS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND " +
-                    "AM_API.API_UUID=AM_GW_API_DEPLOYMENTS.API_ID AND " +
-                    "AM_GW_API_ARTIFACTS.REVISION_ID=AM_GW_API_DEPLOYMENTS.REVISION_ID";
+            "SELECT " +
+                    "   AM_GW_API_DEPLOYMENTS.API_ID AS API_ID," +
+                    "   AM_GW_API_DEPLOYMENTS.REVISION_ID AS REVISION_ID," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_PROVIDER AS API_PROVIDER," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_NAME AS API_NAME," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_VERSION AS API_VERSION," +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_TYPE AS API_TYPE," +
+                    "   AM_GW_API_ARTIFACTS.ARTIFACT AS ARTIFACT," +
+                    "   AM_GW_API_DEPLOYMENTS.LABEL AS LABEL," +
+                    "   AM_GW_API_DEPLOYMENTS.VHOST AS VHOST, " +
+                    "   CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID AS DATA_PLANE_ID, " +
+                    "   AM_API.CONTEXT AS CONTEXT " +
+                    "FROM " +
+                    "   AM_GW_PUBLISHED_API_DETAILS," +
+                    "   AM_GW_API_ARTIFACTS," +
+                    "   AM_GW_API_DEPLOYMENTS," +
+                    "   AM_API," +
+                    "   CHOREO_GW_ENV_DATA_PLANE_MAPPING," +
+                    "   AM_GW_VHOST," +
+                    "   AM_GATEWAY_ENVIRONMENT " +
+                    "WHERE " +
+                    "   CHOREO_GW_ENV_DATA_PLANE_MAPPING.DATA_PLANE_ID IN (_DATA_PLANE_IDS_) AND" +
+                    "   AM_GW_PUBLISHED_API_DETAILS.TENANT_DOMAIN = ? AND" +
+                    "   AM_GW_VHOST.GATEWAY_ENV_ID=AM_GATEWAY_ENVIRONMENT.ID AND" +
+                    "   AM_GATEWAY_ENVIRONMENT.UUID=CHOREO_GW_ENV_DATA_PLANE_MAPPING.GATEWAY_ENV_UUID AND" +
+                    "   AM_GW_VHOST.HOST=AM_GW_API_DEPLOYMENTS.VHOST AND" +
+                    "   AM_GW_PUBLISHED_API_DETAILS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND" +
+                    "   AM_GW_API_ARTIFACTS.API_ID=AM_GW_API_DEPLOYMENTS.API_ID AND" +
+                    "   AM_API.API_UUID=AM_GW_API_DEPLOYMENTS.API_ID AND" +
+                    "   AM_GW_API_ARTIFACTS.REVISION_ID=AM_GW_API_DEPLOYMENTS.REVISION_ID";
 
     public static final String RETRIEVE_ALL_ARTIFACTS_BY_ORGANIZATION_AND_DATA_PLANE_ID =
             "SELECT AM_GW_API_DEPLOYMENTS.API_ID AS API_ID,AM_GW_API_DEPLOYMENTS.REVISION_ID AS REVISION_ID," +
