+++ components/apimgt/org.wso2.carbon.apimgt.internal.service/src/gen/java/org/wso2/carbon/apimgt/internal/service/RetrieveRuntimeMetadataApiService.java
@@ -20,5 +20,5 @@ import javax.ws.rs.core.SecurityContext;
 
 
 public interface RetrieveRuntimeMetadataApiService {
-      public Response retrieveRuntimeMetadataGet(String xWSO2Tenant, String dataPlaneId, MessageContext messageContext) throws APIManagementException;
+      public Response retrieveRuntimeMetadataGet(String dataPlaneId, MessageContext messageContext) throws APIManagementException;
 }
