+++ components/apimgt/org.wso2.carbon.apimgt.internal.service/src/main/resources/api.yaml
@@ -605,7 +605,6 @@ paths:
       description: |
         This will provide access to the deployment metadata for a given data-plane ID in json format
       parameters:
-        - $ref: '#/parameters/requestedTenant'
         - name: dataPlaneId
           in: query
           description: |
